toolkit/components/passwordmgr/test/mochitest/test_insecure_form_field_autocomplete.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test insecure form field autocomplete</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="text/javascript" src="../../../satchel/test/satchel_common.js"></script>
  <script type="text/javascript" src="pwmgr_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>

<script>
var chromeScript = runChecksAfterCommonInit();

var setupScript = runInParent(function setup() {
  const {Services} = ChromeUtils.import("resource://gre/modules/Services.jsm");

  // Create some logins just for this form, since we'll be deleting them.
  var nsLoginInfo = Components.Constructor("@mozilla.org/login-manager/loginInfo;1",
                                           Ci.nsILoginInfo, "init");
  assert.ok(nsLoginInfo != null, "nsLoginInfo constructor");

  let logins = {};
  // login0 has no username, so should be filtered out from the autocomplete list.
  logins.login0 = new nsLoginInfo("http://mochi.test:8888", "http://autocomplete:8888", null,
                                  "", "user0pass", "", "pword");

  logins.login1 = new nsLoginInfo("http://mochi.test:8888", "http://autocomplete:8888", null,
                                  "tempuser1", "temppass1", "uname", "pword");

  logins.login2 = new nsLoginInfo("http://mochi.test:8888", "http://autocomplete:8888", null,
                                  "testuser2", "testpass2", "uname", "pword");

  logins.login3 = new nsLoginInfo("http://mochi.test:8888", "http://autocomplete:8888", null,
                                  "testuser3", "testpass3", "uname", "pword");

  logins.login4 = new nsLoginInfo("http://mochi.test:8888", "http://autocomplete:8888", null,
                                  "zzzuser4", "zzzpass4", "uname", "pword");

  // The following logins are added later
  // login 5 only used in the single-user forms
  logins.login5 = new nsLoginInfo("http://mochi.test:8888", "http://autocomplete2", null,
                                  "singleuser5", "singlepass5", "uname", "pword");

  logins.login6A = new nsLoginInfo("http://mochi.test:8888", "http://autocomplete3", null,
                                   "form7user1", "form7pass1", "uname", "pword");
  logins.login6B = new nsLoginInfo("http://mochi.test:8888", "http://autocomplete3", null,
                                   "form7user2", "form7pass2", "uname", "pword");

  logins.login7  = new nsLoginInfo("http://mochi.test:8888", "http://autocomplete4", null,
                                   "form8user", "form8pass", "uname", "pword");

  logins.login8A = new nsLoginInfo("http://mochi.test:8888", "http://autocomplete5", null,
                                   "form9userAB", "form9pass", "uname", "pword");
  logins.login8B = new nsLoginInfo("http://mochi.test:8888", "http://autocomplete5", null,
                                   "form9userAAB", "form9pass", "uname", "pword");
  logins.login8C = new nsLoginInfo("http://mochi.test:8888", "http://autocomplete5", null,
                                   "form9userAABzz", "form9pass", "uname", "pword");

  logins.login10 = new nsLoginInfo("http://mochi.test:8888", "http://autocomplete7", null,
                                   "testuser10", "testpass10", "uname", "pword");


  // try/catch in case someone runs the tests manually, twice.
  try {
    Services.logins.addLogin(logins.login0);
    Services.logins.addLogin(logins.login1);
    Services.logins.addLogin(logins.login2);
    Services.logins.addLogin(logins.login3);
    Services.logins.addLogin(logins.login4);
  } catch (e) {
    assert.ok(false, "addLogin threw: " + e);
  }

  addMessageListener("addLogin", loginVariableName => {
    let login = logins[loginVariableName];
    assert.ok(!!login, "Login to add is defined: " + loginVariableName);
    Services.logins.addLogin(login);
  });
  addMessageListener("removeLogin", loginVariableName => {
    let login = logins[loginVariableName];
    assert.ok(!!login, "Login to delete is defined: " + loginVariableName);
    Services.logins.removeLogin(login);
  });
});
</script>
<p id="display"></p>

<!-- we presumably can't hide the content for this test. -->
<div id="content">

  <!-- form1 tests multiple matching logins -->
  <form id="form1" action="http://autocomplete:8888/formtest.js" onsubmit="return false;">
    <input  type="text"       name="uname">
    <input  type="password"   name="pword">
    <button type="submit">Submit</button>
  </form>

</div>

<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Login Manager: multiple login autocomplete. **/

SpecialPowers.pushPrefEnv({"set": [["security.allow_eval_with_system_principal", true]]});

var uname = $_(1, "uname");
var pword = $_(1, "pword");

// Restore the form to the default state.
function restoreForm() {
  uname.value = "";
  pword.value = "";
  uname.focus();
}

function sendFakeAutocompleteEvent(element) {
  var acEvent = document.createEvent("HTMLEvents");
  acEvent.initEvent("DOMAutoComplete", true, false);
  element.dispatchEvent(acEvent);
}

function spinEventLoop() {
  return Promise.resolve();
}

add_task(async function setup() {
  await SpecialPowers.pushPrefEnv({"set": [["security.insecure_field_warning.contextual.enabled", true]]});
  listenForUnexpectedPopupShown();
});

add_task(async function test_form1_initial_empty() {
  await SimpleTest.promiseFocus(window);

  // Make sure initial form is empty.
  checkLoginForm(uname, "", pword, "");
  let popupState = await getPopupState();
  is(popupState.open, false, "Check popup is initially closed");
});

add_task(async function test_form1_warning_entry() {
  await SimpleTest.promiseFocus(window);
  // Trigger autocomplete popup
  restoreForm();
  let shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown"); // open
  let results = await shownPromise;

  let popupState = await getPopupState();
  is(popupState.selectedIndex, -1, "Check no entries are selected upon opening");

  let expectedMenuItems = ["This connection is not secure. Logins entered here could be compromised. Learn More",
                           "tempuser1",
                           "testuser2",
                           "testuser3",
                           "zzzuser4"];
  checkAutoCompleteResults(results, expectedMenuItems, "mochi.test", "Check all menuitems are displayed correctly.");

  synthesizeKey("KEY_ArrowDown"); // select insecure warning
  checkLoginForm(uname, "", pword, ""); // value shouldn't update just by selecting
  synthesizeKey("KEY_Enter");
  await spinEventLoop(); // let focus happen
  checkLoginForm(uname, "", pword, "");
});

add_task(async function test_form1_first_entry() {
  await SimpleTest.promiseFocus(window);
  // Trigger autocomplete popup
  restoreForm();
  let shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown"); // open
  await shownPromise;

  let popupState = await getPopupState();
  is(popupState.selectedIndex, -1, "Check no entries are selected upon opening");

  synthesizeKey("KEY_ArrowDown"); // skip insecure warning
  synthesizeKey("KEY_ArrowDown"); // first
  checkLoginForm(uname, "", pword, ""); // value shouldn't update just by selecting
  synthesizeKey("KEY_Enter");
  await promiseFormsProcessed();
  checkLoginForm(uname, "tempuser1", pword, "temppass1");
});

add_task(async function test_form1_second_entry() {
  // Trigger autocomplete popup
  restoreForm();
  let shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown"); // open
  await shownPromise;

  synthesizeKey("KEY_ArrowDown"); // skip insecure warning
  synthesizeKey("KEY_ArrowDown"); // first
  synthesizeKey("KEY_ArrowDown"); // second
  synthesizeKey("KEY_Enter");
  await promiseFormsProcessed();
  checkLoginForm(uname, "testuser2", pword, "testpass2");
});

add_task(async function test_form1_third_entry() {
  // Trigger autocomplete popup
  restoreForm();
  let shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown"); // open
  await shownPromise;

  synthesizeKey("KEY_ArrowDown"); // skip insecure warning
  synthesizeKey("KEY_ArrowDown"); // first
  synthesizeKey("KEY_ArrowDown"); // second
  synthesizeKey("KEY_ArrowDown"); // third
  synthesizeKey("KEY_Enter");
  await promiseFormsProcessed();
  checkLoginForm(uname, "testuser3", pword, "testpass3");
});

add_task(async function test_form1_fourth_entry() {
  // Trigger autocomplete popup
  restoreForm();
  let shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown"); // open
  await shownPromise;

  synthesizeKey("KEY_ArrowDown"); // skip insecure warning
  synthesizeKey("KEY_ArrowDown"); // first
  synthesizeKey("KEY_ArrowDown"); // second
  synthesizeKey("KEY_ArrowDown"); // third
  synthesizeKey("KEY_ArrowDown"); // fourth
  synthesizeKey("KEY_Enter");
  await promiseFormsProcessed();
  checkLoginForm(uname, "zzzuser4", pword, "zzzpass4");
});

add_task(async function test_form1_wraparound_first_entry() {
  // Trigger autocomplete popup
  restoreForm();
  await spinEventLoop(); // let focus happen
  let shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown"); // open
  await shownPromise;

  synthesizeKey("KEY_ArrowDown"); // skip insecure warning
  synthesizeKey("KEY_ArrowDown"); // first
  synthesizeKey("KEY_ArrowDown"); // second
  synthesizeKey("KEY_ArrowDown"); // third
  synthesizeKey("KEY_ArrowDown"); // fourth
  synthesizeKey("KEY_ArrowDown"); // footer
  synthesizeKey("KEY_ArrowDown"); // deselects
  synthesizeKey("KEY_ArrowDown"); // skip insecure warning
  synthesizeKey("KEY_ArrowDown"); // first
  synthesizeKey("KEY_Enter");
  await promiseFormsProcessed();
  checkLoginForm(uname, "tempuser1", pword, "temppass1");
});

add_task(async function test_form1_wraparound_up_last_entry() {
  // Trigger autocomplete popup
  restoreForm();
  let shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown"); // open
  await shownPromise;

  synthesizeKey("KEY_ArrowUp"); // footer
  synthesizeKey("KEY_ArrowUp"); // last (fourth)
  synthesizeKey("KEY_Enter");
  await promiseFormsProcessed();
  checkLoginForm(uname, "zzzuser4", pword, "zzzpass4");
});

add_task(async function test_form1_wraparound_down_up_up() {
  // Trigger autocomplete popup
  restoreForm();
  let shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown"); // open
  await shownPromise;

  synthesizeKey("KEY_ArrowDown"); // select first entry
  synthesizeKey("KEY_ArrowUp"); // selects nothing!
  synthesizeKey("KEY_ArrowUp"); // selects footer
  synthesizeKey("KEY_ArrowUp"); // select last entry
  synthesizeKey("KEY_Enter");
  await promiseFormsProcessed();
  checkLoginForm(uname, "zzzuser4", pword, "zzzpass4");
});

add_task(async function test_form1_wraparound_up_last() {
  restoreForm();
  let shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown"); // open
  await shownPromise;

  synthesizeKey("KEY_ArrowDown");
  synthesizeKey("KEY_ArrowUp"); // deselects
  synthesizeKey("KEY_ArrowUp"); // footer
  synthesizeKey("KEY_ArrowUp"); // last entry
  synthesizeKey("KEY_ArrowUp");
  synthesizeKey("KEY_ArrowUp");
  synthesizeKey("KEY_ArrowUp"); // skip insecure warning
  synthesizeKey("KEY_ArrowUp"); // first entry
  synthesizeKey("KEY_ArrowUp"); // deselects
  synthesizeKey("KEY_ArrowUp"); // footer
  synthesizeKey("KEY_ArrowUp"); // last entry
  synthesizeKey("KEY_Enter");
  await promiseFormsProcessed();
  checkLoginForm(uname, "zzzuser4", pword, "zzzpass4");
});

add_task(async function test_form1_fill_username_without_autofill_right() {
  restoreForm();
  let shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown"); // open
  await shownPromise;

  // Set first entry w/o triggering autocomplete
  synthesizeKey("KEY_ArrowDown"); // skip insecure warning
  synthesizeKey("KEY_ArrowDown"); // first
  synthesizeKey("KEY_ArrowRight");
  await spinEventLoop();
  checkLoginForm(uname, "tempuser1", pword, ""); // empty password
});

add_task(async function test_form1_fill_username_without_autofill_left() {
  restoreForm();
  let shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown"); // open
  await shownPromise;

  // Set first entry w/o triggering autocomplete
  synthesizeKey("KEY_ArrowDown"); // skip insecure warning
  synthesizeKey("KEY_ArrowDown"); // first
  synthesizeKey("KEY_ArrowLeft");
  checkLoginForm(uname, "tempuser1", pword, ""); // empty password
});

add_task(async function test_form1_pageup_first() {
  restoreForm();
  let shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown"); // open
  await shownPromise;

  // Check first entry (page up)
  synthesizeKey("KEY_ArrowDown"); // first
  synthesizeKey("KEY_ArrowDown"); // second
  synthesizeKey("KEY_PageUp"); // first
  synthesizeKey("KEY_ArrowDown"); // skip insecure warning
  synthesizeKey("KEY_Enter");
  await promiseFormsProcessed();
  checkLoginForm(uname, "tempuser1", pword, "temppass1");
});

add_task(async function test_form1_pagedown_last() {
  restoreForm();
  let shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown"); // open
  await shownPromise;

  // test 13
  // Check last entry (page down)
  synthesizeKey("KEY_ArrowDown"); // first
  synthesizeKey("KEY_PageDown"); // last
  synthesizeKey("KEY_ArrowUp"); // skip footer
  synthesizeKey("KEY_Enter");
  await promiseFormsProcessed();
  checkLoginForm(uname, "zzzuser4", pword, "zzzpass4");
});

add_task(async function test_form1_untrusted_event() {
  restoreForm();
  await spinEventLoop();

  // Send a fake (untrusted) event.
  checkLoginForm(uname, "", pword, "");
  uname.value = "zzzuser4";
  sendFakeAutocompleteEvent(uname);
  await spinEventLoop();
  checkLoginForm(uname, "zzzuser4", pword, "");
});

add_task(async function test_form1_delete() {
  restoreForm();
  let shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown"); // open
  await shownPromise;

  // XXX tried sending character "t" before/during dropdown to test
  // filtering, but had no luck. Seemed like the character was getting lost.
  // Setting uname.value didn't seem to work either. This works with a human
  // driver, so I'm not sure what's up.

  synthesizeKey("KEY_ArrowDown"); // skip insecure warning
  // Delete the first entry (of 4), "tempuser1"
  synthesizeKey("KEY_ArrowDown");
  var numLogins;
  numLogins = LoginManager.countLogins("http://mochi.test:8888", "", null);
  is(numLogins, 5, "Correct number of logins before deleting one");

  let countChangedPromise = notifyMenuChanged(5);
  var deletionPromise = promiseStorageChanged(["removeLogin"]);
  // On OS X, shift-backspace and shift-delete work, just delete does not.
  // On Win/Linux, shift-backspace does not work, delete and shift-delete do.
  synthesizeKey("KEY_Delete", {shiftKey: true});
  await deletionPromise;

  checkLoginForm(uname, "", pword, "");
  numLogins = LoginManager.countLogins("http://mochi.test:8888", "", null);
  is(numLogins, 4, "Correct number of logins after deleting one");
  await countChangedPromise;
  synthesizeKey("KEY_Enter");
  await promiseFormsProcessed();
  checkLoginForm(uname, "testuser2", pword, "testpass2");
});

add_task(async function test_form1_first_after_deletion() {
  restoreForm();
  let shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown"); // open
  await shownPromise;

  synthesizeKey("KEY_ArrowDown"); // skip insecure warning
  // Check the new first entry (of 3)
  synthesizeKey("KEY_ArrowDown");
  synthesizeKey("KEY_Enter");
  await promiseFormsProcessed();
  checkLoginForm(uname, "testuser2", pword, "testpass2");
});

add_task(async function test_form1_delete_second() {
  restoreForm();
  let shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown"); // open
  await shownPromise;

  synthesizeKey("KEY_ArrowDown"); // skip insecure warning
  // Delete the second entry (of 3), "testuser3"
  synthesizeKey("KEY_ArrowDown");
  synthesizeKey("KEY_ArrowDown");
  let storageChanged = promiseStorageChanged(["removeLogin"]);
  synthesizeKey("KEY_Delete", {shiftKey: true});
  await storageChanged;
  checkLoginForm(uname, "", pword, "");
  let numLogins = LoginManager.countLogins("http://mochi.test:8888", "", null);
  is(numLogins, 3, "Correct number of logins after deleting one");
  synthesizeKey("KEY_Enter");
  await promiseFormsProcessed();
  checkLoginForm(uname, "zzzuser4", pword, "zzzpass4");
});

add_task(async function test_form1_first_after_deletion2() {
  restoreForm();
  let shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown"); // open
  await shownPromise;

  synthesizeKey("KEY_ArrowDown"); // skip insecure warning
  // Check the new first entry (of 2)
  synthesizeKey("KEY_ArrowDown");
  synthesizeKey("KEY_Enter");
  await promiseFormsProcessed();
  checkLoginForm(uname, "testuser2", pword, "testpass2");
});

add_task(async function test_form1_delete_last() {
  restoreForm();
  let shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown"); // open
  await shownPromise;

  synthesizeKey("KEY_ArrowDown"); // skip insecure warning
  // test 54
  // Delete the last entry (of 2), "zzzuser4"
  synthesizeKey("KEY_ArrowDown");
  synthesizeKey("KEY_ArrowDown");
  let storageChanged = promiseStorageChanged(["removeLogin"]);
  synthesizeKey("KEY_Delete", {shiftKey: true});
  await storageChanged;
  checkLoginForm(uname, "", pword, "");
  let numLogins = LoginManager.countLogins("http://mochi.test:8888", "", null);
  is(numLogins, 2, "Correct number of logins after deleting one");
  synthesizeKey("KEY_ArrowDown"); // skip insecure warning
  synthesizeKey("KEY_ArrowDown");
  synthesizeKey("KEY_Enter");
  await promiseFormsProcessed();
  checkLoginForm(uname, "testuser2", pword, "testpass2");
});

add_task(async function test_form1_first_after_3_deletions() {
  restoreForm();
  let shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown"); // open
  await shownPromise;

  synthesizeKey("KEY_ArrowDown"); // skip insecure warning
  // Check the only remaining entry
  synthesizeKey("KEY_ArrowDown");
  synthesizeKey("KEY_Enter");
  await promiseFormsProcessed();
  checkLoginForm(uname, "testuser2", pword, "testpass2");
});

add_task(async function test_form1_check_only_entry_remaining() {
  restoreForm();
  let shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown"); // open
  await shownPromise;

  synthesizeKey("KEY_ArrowDown"); // skip insecure warning
  // test 56
  // Delete the only remaining entry, "testuser2"
  synthesizeKey("KEY_ArrowDown");
  let storageChanged = promiseStorageChanged(["removeLogin", "removeLogin"]);
  synthesizeKey("KEY_Delete", {shiftKey: true});
  checkLoginForm(uname, "", pword, "");
  let numLogins = LoginManager.countLogins("http://mochi.test:8888", "", null);
  is(numLogins, 1, "Correct number of logins after deleting one");

  // remove the login that's not shown in the list.
  setupScript.sendSyncMessage("removeLogin", "login0");
  await storageChanged;
});

// Tests for single-user forms for ignoring autocomplete=off
add_task(async function test_form2() {
  let storageChanged = promiseStorageChanged(["addLogin"]);
  setupScript.sendSyncMessage("addLogin", "login5");
  await storageChanged;
  await setFormAndWaitForFieldFilled(`
    <!-- other forms test single logins, with autocomplete=off set -->
    <form id="form2" action="http://autocomplete2" onsubmit="return false;">
      <input  type="text"       name="uname">
      <input  type="password"   name="pword" autocomplete="off">
      <button type="submit">Submit</button>
    </form>`, {fieldSelector: `input[name="uname"]`, fieldValue: "singleuser5"});

  // Turn our attention to form2
  uname = $_(2, "uname");
  pword = $_(2, "pword");
  checkLoginForm(uname, "singleuser5", pword, "singlepass5");

  restoreForm();
  let shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown"); // open
  await shownPromise;

  synthesizeKey("KEY_ArrowDown"); // skip insecure warning
  // Check first entry
  synthesizeKey("KEY_ArrowDown");
  checkLoginForm(uname, "", pword, ""); // value shouldn't update
  synthesizeKey("KEY_Enter");
  await promiseFormsProcessed();
  checkLoginForm(uname, "singleuser5", pword, "singlepass5");
});

add_task(async function test_form3() {
  await setFormAndWaitForFieldFilled(`
    <form id="form3" action="http://autocomplete2" onsubmit="return false;">
      <input  type="text"       name="uname" autocomplete="off">
      <input  type="password"   name="pword">
      <button type="submit">Submit</button>
    </form>`, {fieldSelector: `input[name="uname"]`, fieldValue: "singleuser5"});

  uname = $_(3, "uname");
  pword = $_(3, "pword");
  checkLoginForm(uname, "singleuser5", pword, "singlepass5");
  restoreForm();
  let shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown"); // open
  await shownPromise;

  synthesizeKey("KEY_ArrowDown"); // skip insecure warning
  // Check first entry
  synthesizeKey("KEY_ArrowDown");
  checkLoginForm(uname, "", pword, ""); // value shouldn't update
  synthesizeKey("KEY_Enter");
  await promiseFormsProcessed();
  checkLoginForm(uname, "singleuser5", pword, "singlepass5");
});

add_task(async function test_form4() {
  await setFormAndWaitForFieldFilled(`
    <form id="form4" action="http://autocomplete2" onsubmit="return false;" autocomplete="off">
      <input  type="text"       name="uname">
      <input  type="password"   name="pword">
      <button type="submit">Submit</button>
    </form>`, {fieldSelector: `input[name="uname"]`, fieldValue: "singleuser5"});

  uname = $_(4, "uname");
  pword = $_(4, "pword");
  checkLoginForm(uname, "singleuser5", pword, "singlepass5");
  restoreForm();
  let shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown"); // open
  await shownPromise;

  synthesizeKey("KEY_ArrowDown"); // skip insecure warning
  // Check first entry
  synthesizeKey("KEY_ArrowDown");
  checkLoginForm(uname, "", pword, ""); // value shouldn't update
  synthesizeKey("KEY_Enter");
  await promiseFormsProcessed();
  checkLoginForm(uname, "singleuser5", pword, "singlepass5");
});

add_task(async function test_form5() {
  await setFormAndWaitForFieldFilled(`
    <form id="form5" action="http://autocomplete2" onsubmit="return false;">
      <input  type="text"       name="uname" autocomplete="off">
      <input  type="password"   name="pword" autocomplete="off">
      <button type="submit">Submit</button>
    </form>`, {fieldSelector: `input[name="uname"]`, fieldValue: "singleuser5"});

  uname = $_(5, "uname");
  pword = $_(5, "pword");
  checkLoginForm(uname, "singleuser5", pword, "singlepass5");
  restoreForm();
  let shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown"); // open
  await shownPromise;

  synthesizeKey("KEY_ArrowDown"); // skip insecure warning
  // Check first entry
  synthesizeKey("KEY_ArrowDown");
  checkLoginForm(uname, "", pword, ""); // value shouldn't update
  synthesizeKey("KEY_Enter");
  await promiseFormsProcessed();
  checkLoginForm(uname, "singleuser5", pword, "singlepass5");
});

add_task(async function test_form6() {
  await setFormAndWaitForFieldFilled(`
    <!-- control -->
    <form id="form6" action="http://autocomplete2" onsubmit="return false;">
      <input  type="text"       name="uname">
      <input  type="password"   name="pword">
      <button type="submit">Submit</button>
    </form>`, {fieldSelector: `input[name="uname"]`, fieldValue: "singleuser5"});

  // (this is a control, w/o autocomplete=off, to ensure the login
  // that was being suppressed would have been filled in otherwise)
  uname = $_(6, "uname");
  pword = $_(6, "pword");
  checkLoginForm(uname, "singleuser5", pword, "singlepass5");
});

add_task(async function test_form6_changeUsername() {
  // Test that the password field remains filled in after changing
  // the username.
  uname.focus();
  synthesizeKey("KEY_ArrowRight");
  synthesizeKey("X", {shiftKey: true});
  // Trigger the 'blur' event on uname
  pword.focus();
  await spinEventLoop();
  checkLoginForm(uname, "singleuser5X", pword, "singlepass5");
  uname.focus();

  let storageChanged = promiseStorageChanged(["removeLogin"]);
  setupScript.sendSyncMessage("removeLogin", "login5");
  await storageChanged;
});

add_task(async function test_form7() {
  let storageChanged = promiseStorageChanged(["addLogin", "addLogin"]);
  setupScript.sendSyncMessage("addLogin", "login6A");
  setupScript.sendSyncMessage("addLogin", "login6B");
  await storageChanged;
  await setFormAndWaitForFieldFilled(`
    <!-- This form will be manipulated to insert a different username field. -->
    <form id="form7" action="http://autocomplete3" onsubmit="return false;">
      <input  type="text"       name="uname">
      <input  type="password"   name="pword">
      <button type="submit">Submit</button>
    </form>`, {fieldSelector: `input[name="uname"]`, fieldValue: ""});

  uname = $_(7, "uname");
  pword = $_(7, "pword");
  checkLoginForm(uname, "", pword, "");

  // Insert a new username field into the form. We'll then make sure
  // that invoking the autocomplete doesn't try to fill the form.
  var newField = document.createElement("input");
  newField.setAttribute("type", "text");
  newField.setAttribute("name", "uname2");
  pword.parentNode.insertBefore(newField, pword);
  is($_(7, "uname2").value, "", "Verifying empty uname2");

  // Delete login6B. It was created just to prevent filling in a login
  // automatically, removing it makes it more likely that we'll catch a
  // future regression with form filling here.
  storageChanged = promiseStorageChanged(["removeLogin"]);
  setupScript.sendSyncMessage("removeLogin", "login6B");
  await storageChanged;
});

add_task(async function test_form7_2() {
  restoreForm();
  let shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown"); // open
  await shownPromise;

  synthesizeKey("KEY_ArrowDown"); // skip insecure warning
  // Check first entry
  synthesizeKey("KEY_ArrowDown");
  checkLoginForm(uname, "", pword, ""); // value shouldn't update
  synthesizeKey("KEY_Enter");
  // The form changes, so we expect the old username field to get the
  // selected autocomplete value, but neither the new username field nor
  // the password field should have any values filled in.
  await spinEventLoop();
  checkLoginForm(uname, "form7user1", pword, "");
  is($_(7, "uname2").value, "", "Verifying empty uname2");
  restoreForm(); // clear field, so reloading test doesn't fail

  let storageChanged = promiseStorageChanged(["removeLogin"]);
  setupScript.sendSyncMessage("removeLogin", "login6A");
  await storageChanged;
});

add_task(async function test_form8() {
  let storageChanged = promiseStorageChanged(["addLogin"]);
  setupScript.sendSyncMessage("addLogin", "login7");
  await storageChanged;
  await setFormAndWaitForFieldFilled(`
    <!-- This form will be manipulated to insert a different username field. -->
    <form id="form7" action="http://autocomplete3" onsubmit="return false;">
      <input  type="text"       name="uname">
      <input  type="password"   name="pword">
      <button type="submit">Submit</button>
    </form>
    <!-- test for no autofill after onblur with blank username -->
    <form id="form8" action="http://autocomplete4" onsubmit="return false;">
      <input  type="text"       name="uname">
      <input  type="password"   name="pword">
      <button type="submit">Submit</button>
    </form>
    `, {fieldSelector: `#form8 input[name="uname"]`, fieldValue: "form8user"});

  uname = $_(8, "uname");
  pword = $_(8, "pword");
  checkLoginForm(uname, "form8user", pword, "form8pass");
  restoreForm();
});

add_task(async function test_form8_blur() {
  checkLoginForm(uname, "", pword, "");
  // Focus the previous form to trigger a blur.
  $_(7, "uname").focus();
});

add_task(async function test_form8_2() {
  checkLoginForm(uname, "", pword, "");
  restoreForm();
});

add_task(async function test_form8_3() {
  checkLoginForm(uname, "", pword, "");
});

add_task(async function test_form9_filtering() {
  let storageChanged = promiseStorageChanged(["removeLogin", "addLogin", "addLogin"]);
  setupScript.sendSyncMessage("removeLogin", "login7");
  setupScript.sendSyncMessage("addLogin", "login8A");
  setupScript.sendSyncMessage("addLogin", "login8B");
  await storageChanged;
  await setFormAndWaitForFieldFilled(`
    <!-- test autocomplete dropdown -->
    <form id="form9" action="http://autocomplete5" onsubmit="return false;">
      <input  type="text"       name="uname">
      <input  type="password"   name="pword">
      <button type="submit">Submit</button>
    </form>`, {fieldSelector: `input[name="uname"]`, fieldValue: ""});

  // Turn our attention to form9 to test the dropdown - bug 497541
  let shownPromise = promiseACShown();
  uname = $_(9, "uname");
  pword = $_(9, "pword");
  uname.focus();
  let results = await shownPromise;
  checkAutoCompleteResults(results,
                           ["This connection is not secure. Logins entered here could be compromised. Learn More", "form9userAAB", "form9userAB"],
                           "mochi.test",
                           "Check dropdown is showing all logins while field is blank");
  synthesizeKey("KEY_Escape"); // Need to close the popup so we can get another popupshown after sending the string below.
  shownPromise = promiseACShown();
  sendString("form9userAB");
  results = await shownPromise;
  checkAutoCompleteResults(results,
                           ["This connection is not secure. Logins entered here could be compromised. Learn More", "form9userAB"],
                           "mochi.test",
                           "Check dropdown is showing login with only one 'A'");

  checkLoginForm(uname, "form9userAB", pword, "");
  uname.focus();
  synthesizeKey("KEY_ArrowLeft");
  shownPromise = promiseACShown();
  synthesizeKey("A", {shiftKey: true});
  results = await shownPromise;

  checkLoginForm(uname, "form9userAAB", pword, "");
  checkAutoCompleteResults(results,
                           ["This connection is not secure. Logins entered here could be compromised. Learn More", "form9userAAB"],
                           "mochi.test",
                           "Check dropdown is updated after inserting 'A'");
  synthesizeKey("KEY_ArrowDown"); // skip insecure warning
  synthesizeKey("KEY_ArrowDown");
  synthesizeKey("KEY_Enter");
  await promiseFormsProcessed();
  checkLoginForm(uname, "form9userAAB", pword, "form9pass");
});

add_task(async function test_form9_autocomplete_cache() {
  // Note that this addLogin call will only be seen by the autocomplete
  // attempt for the synthesizeKey if we do not successfully cache the
  // autocomplete results.
  let storageChanged = promiseStorageChanged(["addLogin"]);
  setupScript.sendSyncMessage("addLogin", "login8C");
  await storageChanged;
  uname.focus();
  let promise2 = notifyMenuChanged(2);
  let shownPromise = promiseACShown();
  sendString("z");
  let results = await promise2;
  await shownPromise;
  checkAutoCompleteResults(results,
                           ["This connection is not secure. Logins entered here could be compromised. Learn More"],
                           "mochi.test",
                           "Check popup only has the footer and insecure warning");
  // check that empty results are cached - bug 496466
  promise2 = notifyMenuChanged(2);
  sendString("z");
  results = await promise2;
  checkAutoCompleteResults(results,
                           ["This connection is not secure. Logins entered here could be compromised. Learn More"],
                           "mochi.test",
                           "Check popup only has the footer and insecure warning");
  storageChanged = promiseStorageChanged(["removeLogin", "removeLogin", "removeLogin"]);
  setupScript.sendSyncMessage("removeLogin", "login8A");
  setupScript.sendSyncMessage("removeLogin", "login8B");
  setupScript.sendSyncMessage("removeLogin", "login8C");
  await storageChanged;
});

add_task(async function test_form11_recipes() {
  let storageChanged = promiseStorageChanged(["addLogin"]);
  setupScript.sendSyncMessage("addLogin", "login10");
  await storageChanged;
  await loadRecipes({
    siteRecipes: [{
      "hosts": ["mochi.test:8888"],
      "usernameSelector": "input[name='1']",
      "passwordSelector": "input[name='2']",
    }],
  });
  await setFormAndWaitForFieldFilled(`
    <!-- test for onUsernameInput recipe testing -->
    <form id="form11" action="http://autocomplete7" onsubmit="return false;">
      <input  type="text"   name="1">
      <input  type="text"   name="2">
      <button type="submit">Submit</button>
    </form>`, {fieldSelector: `input[name="1"]`, fieldValue: ""});

  uname = $_(11, "1");
  pword = $_(11, "2");

  // First test DOMAutocomplete
  // Switch the password field to type=password so _fillForm marks the username
  // field for autocomplete.
  pword.type = "password";
  await promiseFormsProcessed();
  restoreForm();
  checkLoginForm(uname, "", pword, "");
  let shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown"); // open
  await shownPromise;

  synthesizeKey("KEY_ArrowDown"); // skip insecure warning
  synthesizeKey("KEY_ArrowDown");
  checkLoginForm(uname, "", pword, ""); // value shouldn't update
  synthesizeKey("KEY_Enter");
  await promiseFormsProcessed();
  checkLoginForm(uname, "testuser10", pword, "testpass10");

  // Now test recipes with blur on the username field.
  restoreForm();
  checkLoginForm(uname, "", pword, "");
  uname.value = "testuser10";
  checkLoginForm(uname, "testuser10", pword, "");
  synthesizeKey("KEY_Tab");
  await promiseFormsProcessed();
  checkLoginForm(uname, "testuser10", pword, "testpass10");
  await resetRecipes();
});

add_task(async function test_form12_formless() {
  // Test form-less autocomplete
  await setFormAndWaitForFieldFilled(`
    <!-- tests <form>-less autocomplete -->
    <div id="form12">
      <input  type="text"       name="uname" id="uname">
      <input  type="password"   name="pword" id="pword">
      <button type="submit">Submit</button>
    </div>`, {fieldSelector: `input[name="uname"]`, fieldValue: ""});

  uname = $_(12, "uname");
  pword = $_(12, "pword");
  let shownPromise = promiseACShown();
  uname.focus();
  await shownPromise;
  synthesizeKey("KEY_Escape"); // close the autocomplete popup
  shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown"); // open
  let results = await shownPromise;

  let expectedMenuItems = ["This connection is not secure. Logins entered here could be compromised. Learn More",
                           "testuser10"];
  checkAutoCompleteResults(results, expectedMenuItems, "mochi.test", "Check all menuitems are displayed correctly.");
  synthesizeKey("KEY_ArrowDown"); // skip insecure warning
  synthesizeKey("KEY_ArrowDown");
  checkLoginForm(uname, "", pword, ""); // value shouldn't update
  let processedPromise = promiseFormsProcessed();
  synthesizeKey("KEY_Enter");
  await processedPromise;
  checkLoginForm(uname, "testuser10", pword, "testpass10");
});

add_task(async function test_form13_stays_open_upon_empty_search() {
  await setFormAndWaitForFieldFilled(`
    <!-- test not closing when the search string (.value) becomes empty -->
    <form id="form13" action="http://autocomplete:8888/formtest.js" onsubmit="return false;">
      <input  type="text"       name="uname" value="prefilled">
      <input  type="password"   name="pword" value="prefilled">
      <button type="submit">Submit</button>
    </form>`, {fieldSelector: `input[name="uname"]`, fieldValue: "prefilled"});

  uname = $_(13, "uname");
  pword = $_(13, "pword");
  checkLoginForm(uname, "prefilled", pword, "prefilled");

  uname.scrollIntoView();
  let shownPromise = promiseACShown();
  synthesizeMouseAtCenter(uname, {});
  await shownPromise;
  uname.select();
  synthesizeKey("KEY_Delete");

  await spinEventLoop();
  let popupState = await getPopupState();
  is(popupState.open, true, "Check popup is still open");
  checkLoginForm(uname, "", pword, "prefilled");

  info("testing password field");
  synthesizeMouseAtCenter(pword, {});
  pword.select();
  popupState = await getPopupState();
  is(popupState.open, false, "Check popup closed since password field isn't empty");
  shownPromise = promiseACShown();
  synthesizeKey("KEY_Delete");
  await shownPromise;
  checkLoginForm(uname, "", pword, "");
});
</script>
</pre>
</body>
</html>
